# agent/design/api-contract.yaml
openapi: 3.0.3
info:
  title: PumpBuddy API
  version: 0.1.0
  description: |
    Canonical API contract for the active workout product slice.
    Note: the 2026-03-28 auth/session boundary refactor is internal-only and does
    not change external endpoints or payload schemas.
    Note: the 2026-04-06 workout-select greeting update reuses existing
    AuthSessionUser.display_name and does not change endpoint shapes.
    Note: the 2026-04-12 auth-failure logging hardening emits structured
    `auth_login_failure` telemetry for application attempt keys without
    endpoint or payload-schema changes.
    Note: the 2026-04-12 settings profile data update extends AuthSessionUser
    with `login` and `registration_date` for settings display.
    Note: the 2026-04-12 settings display-name persistence update adds PATCH
    /auth/session for authenticated display-name changes.
    Note: the 2026-04-13 settings preferences contract update extends PATCH
    /auth/session and AuthSessionUser with `favorite_gym_id`, and adds
    POST /auth/password for authenticated password changes.
    Note: the 2026-05-30 password-change lifecycle hardening keeps
    /auth/password request and response schemas unchanged while throttling
    repeated current-password failures and clearing the session cookie after a
    successful password change.
    Note: the 2026-04-14 About navigation wiring tightened renderer
    AboutMetadata.channel typing to `stable`; this is API-surface alignment
...
paths:
  /auth/login:
    post:
      summary: Log in with login and password
      operationId: loginWithCredentials
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthLoginRequest'
      responses:
        '200':
          description: Login succeeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthLoginResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /auth/session:
    get:
      summary: Resolve session state
      operationId: resolveSession
      responses:
        '200':
          description: Session resolved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSessionResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
    patch:
      summary: Update authenticated session user settings fields
      operationId: updateSessionProfile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthUpdateDisplayNameRequest'
      responses:
        '200':
          description: Session profile updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSessionResponse'
        '400':
          description: Invalid profile update payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /auth/session/side-menu-middle-clicks:
    post:
      summary: Increment authenticated side-menu middle click count
      operationId: incrementSideMenuMiddleClickCount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthIncrementSideMenuMiddleClickRequest'
      responses:
        '200':
          description: Side-menu middle click count incremented
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSessionResponse'
        '400':
          description: Invalid side-menu preference update payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /auth/password:
    post:
      summary: Update authenticated user password
      operationId: updatePassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthUpdatePasswordRequest'
      responses:
        '204':
          description: Password updated successfully and session cookie cleared
        '400':
          description: Invalid password update payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: 'Current password validation failed (message: "Current password validation failed")'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /auth/logout:
    post:
      summary: Log out current session
      operationId: logoutSession
      responses:
        '204':
          description: Logout succeeded
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/about:
    get:
      summary: Retrieve About metadata for renderer display
      operationId: getAboutMetadata
      responses:
        '200':
          description: About metadata retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AboutMetadataResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/gyms:
    get:
      summary: List available gyms
      operationId: listGyms
      responses:
        '200':
          description: |
            Gyms retrieved successfully for the authenticated user. Rows are
            ordered with the user's configured favorite gym first when visible,
            followed by gym display name with stable backend tie-breakers.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GymSummary'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      summary: Create a draft gym
      operationId: createGym
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GymCreateRequest'
      responses:
        '201':
          description: Draft gym created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GymSummary'
        '400':
          description: Invalid gym create payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Duplicate normalized gym name
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/load-profiles:
    get:
      summary: List available load profiles
      operationId: listLoadProfiles
      responses:
        '200':
          description: |
            Load profiles retrieved successfully for the authenticated user.
            Rows are ordered alphabetically by display name with inactive
            profiles grouped last and stable backend tie-breakers.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LoadProfileSummary'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      summary: Create a draft load profile
      operationId: createLoadProfile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoadProfileCreateRequest'
      responses:
        '201':
          description: Draft load profile created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoadProfileSummary'
        '400':
          description: Invalid load profile create payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Duplicate load profile name or invalid durable state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/load-profiles/{loadProfileId}:
    get:
      summary: Load one load profile detail
      operationId: getLoadProfile
      parameters:
        - name: loadProfileId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Load profile detail retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoadProfileDetailResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Load profile not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
    patch:
      summary: Update a load profile
      operationId: updateLoadProfile
      parameters:
        - name: loadProfileId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoadProfileUpdateRequest'
      responses:
        '200':
          description: Load profile updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoadProfileSummary'
        '400':
          description: Invalid load profile update payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Load profile not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Duplicate load profile name or immutable field update
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      summary: Delete a draft load profile
      operationId: deleteLoadProfile
      parameters:
        - name: loadProfileId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Draft load profile deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Load profile not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Load profile cannot be deleted in its current state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/gyms/{gymId}:
    get:
      summary: Retrieve read-only gym transparency detail
      operationId: getGymDetail
      parameters:
        - name: gymId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Gym detail retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GymDetailResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Gym not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
    patch:
      summary: Rename a gym
      operationId: updateGym
      parameters:
        - name: gymId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GymUpdateRequest'
      responses:
        '200':
          description: Gym renamed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GymSummary'
        '400':
          description: Invalid gym update payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Gym not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Duplicate normalized gym name
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      summary: Delete a draft gym
      operationId: deleteGym
      parameters:
        - name: gymId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Draft gym deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Gym not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Gym cannot be deleted in its current state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/gyms/{gymId}/stations/{stationId}:
    get:
      summary: Retrieve read-only station transparency detail
      operationId: getGymStationDetail
      parameters:
        - name: gymId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: stationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Station detail retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GymStationDetailResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Gym station not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/training-plans:
    get:
      summary: List available training plans
      operationId: listTrainingPlans
      responses:
        '200':
          description: Training plans retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TrainingPlanSummary'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  responses:
    Unauthorized:
      description: Unauthorized (empty response body)
...
  schemas:
...
    AuthLoginRequest:
      type: object
      required:
        - login
        - password
      properties:
        login:
          type: string
          minLength: 1
          description: Explicit public login identifier. Blank values are invalid.
        password:
          type: string
    AuthLoginResponse:
      type: object
      required:
        - authenticated
      properties:
        authenticated:
          type: boolean
...