# backend/src/api/router.rs
use axum::{
    extract::{rejection::JsonRejection, Extension, Path, Query, State},
    routing::{delete, get, post, put},
    Json, Router,
};

// `WorkoutValidationError` helper moved to handlers module; router doesn't need it.

use super::handlers::{
    cancel_active_workout, complete_active_workout, confirm_active_workout_set,
    create_active_workout, create_configurator_station, create_gym, create_load_profile,
    create_workout, delete_configurator_station, delete_gym, delete_latest_active_workout_set,
    delete_load_profile, get_about_metadata, get_active_workout, get_configurator_station,
    get_gym_detail, get_gym_station_detail, get_load_profile, get_training_plan,
    get_workout_detail, get_workout_exercises_performance, get_workout_progress,
    get_workout_summary, list_configurator_stations, list_gyms, list_load_profiles,
    list_training_plan_exercise_variants, list_training_plans, list_workouts,
    reopen_active_workout_exercise, select_active_workout_exercise_option,
    skip_active_workout_exercise, update_active_workout, update_configurator_station, update_gym,
    update_load_profile,
};

use super::middleware;
use super::models::{
    CompleteActiveWorkoutRequest, ConfiguratorStationCreateRequest,
    ConfiguratorStationUpdateRequest, ConfirmActiveWorkoutSetRequest, CreateActiveWorkoutRequest,
    CreateWorkoutRequest, GymWriteRequest, ReopenActiveWorkoutExerciseRequest,
    SelectActiveWorkoutExerciseOptionRequest, SkipActiveWorkoutExerciseRequest,
    TrainingPlanDetailQuery, TrainingPlanExerciseVariantsQuery, UpdateActiveWorkoutRequest,
};
use super::session::AuthenticatedSession;
use super::AppState;
// router does not need persistence error mapping or ApiError directly
pub fn app_router(app_state: AppState) -> Router {
    let api = Router::new()
        .route(
            "/about",
            get(
                |State(state): State<AppState>,
                 Extension(session): Extension<AuthenticatedSession>| async move {
                    get_about_metadata(State(state), Extension(session)).await
                },
            ),
        )
        .route(
            "/gyms",
            get(
                |State(state): State<AppState>,
                 Extension(session): Extension<AuthenticatedSession>| async move {
                    list_gyms(State(state), Extension(session)).await
                },
            )
            .post(
                |State(state): State<AppState>,
                 Extension(session): Extension<AuthenticatedSession>,
                 payload: Result<Json<GymWriteRequest>, JsonRejection>| async move {
                    create_gym(State(state), Extension(session), payload).await
                },
            ),
        )
        .route(
            "/gyms/{gym_id}",
            get(
                |State(state): State<AppState>,
                 Extension(session): Extension<AuthenticatedSession>,
                 Path(gym_id): Path<String>| async move {
                    get_gym_detail(State(state), Extension(session), Path(gym_id)).await
                },
            )
            .patch(
                |State(state): State<AppState>,
                 Extension(session): Extension<AuthenticatedSession>,
                 Path(gym_id): Path<String>,
                 payload: Result<Json<GymWriteRequest>, JsonRejection>| async move {
                    update_gym(State(state), Extension(session), Path(gym_id), payload).await
                },
            )
            .delete(
                |State(state): State<AppState>,
                 Extension(session): Extension<AuthenticatedSession>,
                 Path(gym_id): Path<String>| async move {
                    delete_gym(State(state), Extension(session), Path(gym_id)).await
                },
            ),
        )
        .route(
            "/gyms/{gym_id}/stations/{station_id}",
            get(
                |State(state): State<AppState>,
                 Extension(session): Extension<AuthenticatedSession>,
                 Path((gym_id, station_id)): Path<(String, String)>| async move {
                    get_gym_station_detail(
                        State(state),
                        Extension(session),
                        Path((gym_id, station_id)),
                    )
                    .await
                },
            ),
        )
        .route(
            "/gyms/{gym_id}/configurator-stations",
            get(|State(state): State<AppState>, Extension(session): Extension<AuthenticatedSession>, Path(gym_id): Path<String>| async move { list_configurator_stations(State(state), Extension(session), Path(gym_id)).await })
                .post(|State(state): State<AppState>, Extension(session): Extension<AuthenticatedSession>, Path(gym_id): Path<String>, payload: Result<Json<ConfiguratorStationCreateRequest>, JsonRejection>| async move { create_configurator_station(State(state), Extension(session), Path(gym_id), payload).await }),
        )
        .route(
            "/gyms/{gym_id}/configurator-stations/{station_id}",
            get(|State(state): State<AppState>, Extension(session): Extension<AuthenticatedSession>, Path(ids): Path<(String, String)>| async move { get_configurator_station(State(state), Extension(session), Path(ids)).await })
                .patch(|State(state): State<AppState>, Extension(session): Extension<AuthenticatedSession>, Path(ids): Path<(String, String)>, payload: Result<Json<ConfiguratorStationUpdateRequest>, JsonRejection>| async move { update_configurator_station(State(state), Extension(session), Path(ids), payload).await })
                .delete(|State(state): State<AppState>, Extension(session): Extension<AuthenticatedSession>, Path(ids): Path<(String, String)>| async move { delete_configurator_station(State(state), Extension(session), Path(ids)).await }),
        )
        .route(
            "/load-profiles",
            get(
                |State(state): State<AppState>,
                 Extension(session): Extension<AuthenticatedSession>| async move {
                    list_load_profiles(State(state), Extension(session)).await
                },
            )
            .post(
                |State(state): State<AppState>,
                 Extension(session): Extension<AuthenticatedSession>,
                 Json(payload): Json<super::models::LoadProfileCreateRequest>| async move {
                    create_load_profile(State(state), Extension(session), Json(payload)).await
                },
            ),
        )
        .route(
            "/load-profiles/{load_profile_id}",
            get(
                |State(state): State<AppState>,
                 Extension(session): Extension<AuthenticatedSession>,
                 Path(load_profile_id): Path<String>| async move {
                    get_load_profile(State(state), Extension(session), Path(load_profile_id)).await
                },
            )
            .patch(
                |State(state): State<AppState>,
                 Extension(session): Extension<AuthenticatedSession>,
                 Path(load_profile_id): Path<String>,
                 Json(payload): Json<super::models::LoadProfileUpdateRequest>| async move {
                    update_load_profile(
                        State(state),
                        Extension(session),
                        Path(load_profile_id),
                        Json(payload),
                    )
                    .await
                },
            )
            .delete(
                |State(state): State<AppState>,
                 Extension(session): Extension<AuthenticatedSession>,
                 Path(load_profile_id): Path<String>| async move {
                    delete_load_profile(State(state), Extension(session), Path(load_profile_id))
                        .await
                },
            ),
        )
        .route(
            "/training-plans",
            get(
                |State(state): State<AppState>,
                 Extension(session): Extension<AuthenticatedSession>| async move {
                    list_training_plans(State(state), Extension(session)).await
                },
            ),
        )
        .route(
            "/training-plans/{training_plan_id}",
            get(
                |State(state): State<AppState>,
                 Extension(session): Extension<AuthenticatedSession>,
                 Path(training_plan_id): Path<String>,
                 Query(query): Query<TrainingPlanDetailQuery>| async move {
                    get_training_plan(
                        State(state),
                        Extension(session),
                        Path(training_plan_id),
                        Query(query),
                    )
                    .await
                },
            ),
        )
        .route(
            "/training-plans/{training_plan_id}/options",
            get(
                |State(state): State<AppState>,
                 Extension(session): Extension<AuthenticatedSession>,
                 Path(training_plan_id): Path<String>,
                 Query(query): Query<TrainingPlanExerciseVariantsQuery>| async move {
                    list_training_plan_exercise_variants(
                        State(state),
                        Extension(session),
                        Path(training_plan_id),
                        Query(query),
                    )
                    .await
                },
            ),
        )
        .route(
            "/workouts",
            get(
                |State(state): State<AppState>,
                 Extension(session): Extension<AuthenticatedSession>| async move {
                    list_workouts(State(state), Extension(session)).await
                },
            )
            .post(
                |State(state): State<AppState>,
                 Extension(session): Extension<AuthenticatedSession>,
                 Json(payload): Json<CreateWorkoutRequest>| async move {
                    create_workout(State(state), Extension(session), Json(payload)).await
                },
            ),
        )
        .route(
            "/workouts/progress",
            get(
                |State(state): State<AppState>,
                 Extension(session): Extension<AuthenticatedSession>| async move {
                    get_workout_progress(State(state), Extension(session)).await
                },
            ),
        )
        .route(
            "/workouts/exercises-performance",
            get(
                |State(state): State<AppState>,
                 Extension(session): Extension<AuthenticatedSession>| async move {
                    get_workout_exercises_performance(State(state), Extension(session)).await
                },
            ),
        )
        .route(
            "/active-workout",
            get(
                |State(state): State<AppState>,
                 Extension(session): Extension<AuthenticatedSession>| async move {
                    get_active_workout(State(state), Extension(session)).await
                },
            )
            .post(
                |State(state): State<AppState>,
                 Extension(session): Extension<AuthenticatedSession>,
                 Json(payload): Json<CreateActiveWorkoutRequest>| async move {
                    create_active_workout(State(state), Extension(session), Json(payload)).await
                },
            ),
        )
        .route(
            "/active-workout/{workout_id}",
            put(
                |State(state): State<AppState>,
                 Extension(session): Extension<AuthenticatedSession>,
                 Path(workout_id): Path<String>,
                 Json(payload): Json<UpdateActiveWorkoutRequest>| async move {
                    update_active_workout(
                        State(state),
                        Extension(session),
                        Path(workout_id),
                        Json(payload),
                    )
                    .await
                },
            )
            .delete(
                |State(state): State<AppState>,
                 Extension(session): Extension<AuthenticatedSession>,
                 Path(workout_id): Path<String>| async move {
                    cancel_active_workout(State(state), Extension(session), Path(workout_id)).await
                },
            ),
        )
        .route(
            "/active-workout/{workout_id}/exercises/{exercise_position}/option",
            post(
                |State(state): State<AppState>,
                 Extension(session): Extension<AuthenticatedSession>,
                 Path((workout_id, exercise_position)): Path<(String, i32)>,
                 Json(payload): Json<SelectActiveWorkoutExerciseOptionRequest>| async move {
                    select_active_workout_exercise_option(
                        State(state),
                        Extension(session),
                        Path((workout_id, exercise_position)),
                        Json(payload),
                    )
                    .await
                },
            ),
        )
        .route(
            "/active-workout/{workout_id}/exercises/{exercise_position}/sets",
            post(
                |State(state): State<AppState>,
                 Extension(session): Extension<AuthenticatedSession>,
                 Path((workout_id, exercise_position)): Path<(String, i32)>,
                 Json(payload): Json<ConfirmActiveWorkoutSetRequest>| async move {
                    confirm_active_workout_set(
                        State(state),
                        Extension(session),
                        Path((workout_id, exercise_position)),
                        Json(payload),
                    )
                    .await
                },
            ),
        )
        .route(
            "/active-workout/{workout_id}/exercises/{exercise_position}/sets/latest",
            delete(
                |State(state): State<AppState>,
                 Extension(session): Extension<AuthenticatedSession>,
                 Path((workout_id, exercise_position)): Path<(String, i32)>| async move {
                    delete_latest_active_workout_set(
                        State(state),
                        Extension(session),
                        Path((workout_id, exercise_position)),
                    )
                    .await
                },
            ),
        )
        .route(
            "/active-workout/{workout_id}/exercises/{exercise_position}/skip",
            post(
                |State(state): State<AppState>,
                 Extension(session): Extension<AuthenticatedSession>,
                 Path((workout_id, exercise_position)): Path<(String, i32)>,
                 Json(payload): Json<SkipActiveWorkoutExerciseRequest>| async move {
                    skip_active_workout_exercise(
                        State(state),
                        Extension(session),
                        Path((workout_id, exercise_position)),
                        Json(payload),
                    )
                    .await
                },
            ),
        )
        .route(
            "/active-workout/{workout_id}/reopen",
            post(
                |State(state): State<AppState>,
                 Extension(session): Extension<AuthenticatedSession>,
                 Path(workout_id): Path<String>,
                 Json(payload): Json<ReopenActiveWorkoutExerciseRequest>| async move {
                    reopen_active_workout_exercise(
                        State(state),
                        Extension(session),
                        Path(workout_id),
                        Json(payload),
                    )
                    .await
                },
            ),
        )
        .route(
            "/active-workout/{workout_id}/complete",
            post(
                |State(state): State<AppState>,
                 Extension(session): Extension<AuthenticatedSession>,
                 Path(workout_id): Path<String>,
                 Json(payload): Json<CompleteActiveWorkoutRequest>| async move {
                    complete_active_workout(
                        State(state),
                        Extension(session),
                        Path(workout_id),
                        Json(payload),
                    )
                    .await
                },
            ),
        )
        .route(
            "/workouts/{workout_id}/summary",
            get(
                |State(state): State<AppState>,
                 Extension(session): Extension<AuthenticatedSession>,
                 Path(workout_id): Path<String>| async move {
                    get_workout_summary(State(state), Extension(session), Path(workout_id)).await
                },
            ),
        )
        .route(
            "/workouts/{workout_id}",
            get(
                |State(state): State<AppState>,
                 Extension(session): Extension<AuthenticatedSession>,
                 Path(workout_id): Path<String>| async move {
                    get_workout_detail(State(state), Extension(session), Path(workout_id)).await
                },
            ),
        );

    Router::new()
        .route("/health", get(|| async { "ok" }))
        .route("/auth/login", post(super::auth::login))
        .route("/auth/password", post(super::auth::update_password_handler))
        .route("/auth/logout", post(super::auth::logout))
        .route(
            "/auth/session",
            get(super::auth::session).patch(super::auth::update_session),
        )
        .route(
            "/auth/session/side-menu-middle-clicks",
            post(super::auth::increment_side_menu_middle_click),
        )
        .nest(
            "/api",
            api.layer(axum::middleware::from_fn_with_state(
                app_state.clone(),
                middleware::require_session,
            )),
        )
        .with_state(app_state)
}